- #075_fix_피드백 반영
튜터링 시간 및 슬랙에서 받은 피드백들을 반영했습니다.
무한 스크롤 및 로드 기능까지 추가하였습니다.
시간 제한을 1분으로 설정했습니다.

+++ app/src/main/java/team_h/boostcamp/myapplication/data/repository/DiaryRepository.java
@@ -1,5 +1,7 @@
 package team_h.boostcamp.myapplication.data.repository;
 
+import android.util.Log;
+
 import java.util.List;
 
 import androidx.annotation.NonNull;
@@ -9,8 +11,8 @@ import io.reactivex.schedulers.Schedulers;
 import team_h.boostcamp.myapplication.data.remote.deepaffects.AnalyzedEmotionMapper;
 import team_h.boostcamp.myapplication.data.remote.deepaffects.DeepAffectApiClient;
 import team_h.boostcamp.myapplication.data.remote.deepaffects.request.EmotionAnalyzeRequest;
+import team_h.boostcamp.myapplication.data.room.dao.DiaryDao;
 import team_h.boostcamp.myapplication.model.Diary;
-import team_h.boostcamp.myapplication.model.source.local.DiaryDao;
 
 
 /*
@@ -41,7 +43,7 @@ public class DiaryRepository implements DiaryRepositoryContract {
 
     @NonNull
     @Override
-    public Single<List<Diary>> loadMoreDiaryItems(final int idx) {
+    public Single<List<Diary>> loadMoreDiaryItems(final long idx) {
         return diaryDao.loadMoreDiary(idx).subscribeOn(Schedulers.io());
     }
